requirements.txt
fastapi
uvicorn
neo4j
weaviate-client>=4.0.0
celery
redis
autogen
requests
pytest
